@@ -3,23 +3,21 @@ description: Calendar planning application.
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
+  flutter_localizations:
+    sdk: flutter 
   cupertino_icons: ^1.0.2
-  firebase_core: ^1.0.1
-  firebase_auth: ^1.0.1
-  firebase_auth_web: ^1.0.2
-  modal_progress_hud: ^0.1.3
-  calendar_views: ^0.5.2
+  firebase_core: ^1.3.0
+  firebase_auth: ^1.4.1
   intl: ^0.17.0
   scroll_snap_list: ^0.7.0
-  firebase_database: ^6.1.0
-  tuple: ^1.0.3
-  persistent_bottom_nav_bar: any
-  flutter_datetime_picker: ^1.5.0
+  firebase_database: ^7.1.1
+  tuple: ^2.0.0
+  persistent_bottom_nav_bar: ^4.0.2
   google_fonts: ^2.0.0
   loader_overlay: ^2.0.0
   rxdart: ^0.26.0
@@ -28,9 +26,14 @@ dependencies:
   image_picker: ^0.7.4
   uuid: ^3.0.3
   firebase_messaging: ^9.1.1
-  provider: ^4.3.3
+  provider: ^5.0.0
   maps_launcher: ^2.0.0
   collection: ^1.15.0
+  flutter_svg: ^0.22.0
+  snapping_sheet: ^3.1.0
+  quiver: ^3.0.1
+  url_launcher: ^6.0.6
+  flutter_icons: ^1.1.0
 
 dev_dependencies:
   flutter_test:
