@@ -8,10 +8,10 @@ import 'package:firebase_database/firebase_database.dart' as fb;
 
 class FirebaseOutlaysManager {
   String company;
-  fb.DatabaseReference database;
-  fb.DatabaseReference outlaysRef;
+  late fb.DatabaseReference database;
+  late fb.DatabaseReference outlaysRef;
 
-  FirebaseOutlaysManager({@required this.company}) {
+  FirebaseOutlaysManager({required this.company}) {
     database = fb.FirebaseDatabase.instance.reference();
     outlaysRef = database.reference().child(company).child('Outlays');
   }
