@@ -1,17 +1,19 @@
-import 'package:firebase_core/firebase_core.dart';
 import 'package:firebase_database/firebase_database.dart';
-import 'package:flutter/material.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import '../model/user_parameters.dart';
 
 class FirebaseUserManager {
-  String userToken;
-  User user;
+  String? userToken;
+  User? user;
   FirebaseUserManager();
 
-  Future<UserParameters> getUser() {
-    user = FirebaseAuth.instance.currentUser;
-    return _getUserParameters(user.uid);
+  Future<UserParameters>? getUser() {
+    User? user = FirebaseAuth.instance.currentUser;
+    if (user == null) {
+      return null;
+    } else {
+      return _getUserParameters(user.uid);
+    }
   }
 
   Future<UserParameters> _getUserParameters(String userToken) async {
@@ -24,14 +26,9 @@ class FirebaseUserManager {
         company: snapshot.value['company'],
         isAdmin: snapshot.value['isAdmin'],
         token: userToken,
-        email: user.email,
+        email: user?.email ?? "",
         name: snapshot.value['name'],
         profilePicturePath: snapshot.value['profilePicturePath']);
-  }
-
-  String _getCurrentUserToken() {
-    user = FirebaseAuth.instance.currentUser;
-    return user.uid;
   }
 
   Future<void> setUserProfileImage(UserParameters user, String filePath) async {
