@@ -6,7 +6,7 @@ class FutureImageListWidget extends StatelessWidget {
   final List<String> paths;
   final FirebaseStorageManager firebaseStorageManager;
   FutureImageListWidget(
-      {Key key, @required this.paths, @required this.firebaseStorageManager})
+      {Key? key, required this.paths, required this.firebaseStorageManager})
       : super(key: key);
   @override
   Widget build(BuildContext context) {
@@ -33,7 +33,7 @@ class FutureImageWidget extends StatefulWidget {
   final String path;
   final FirebaseStorageManager firebaseStorageManager;
   FutureImageWidget(
-      {Key key, @required this.path, @required this.firebaseStorageManager})
+      {Key? key, required this.path, required this.firebaseStorageManager})
       : super(key: key);
   @override
   _FutureImageWidgetState createState() => _FutureImageWidgetState();
@@ -43,7 +43,7 @@ class _FutureImageWidgetState extends State<FutureImageWidget>
     with AutomaticKeepAliveClientMixin<FutureImageWidget> {
   @override
   bool get wantKeepAlive => true;
-  Future<Image> _imageFuture;
+  late Future<Image?> _imageFuture;
 
   @override
   void initState() {
@@ -58,13 +58,14 @@ class _FutureImageWidgetState extends State<FutureImageWidget>
       key: GlobalKey(),
       child: FutureBuilder(
           future: _imageFuture,
-          builder: (context, AsyncSnapshot<Image> snapshot) {
+          builder: (context, AsyncSnapshot<Image?> snapshot) {
+            if (snapshot.data == null) return Container();
             if (snapshot.connectionState == ConnectionState.done) {
               return GestureDetector(
                 onTap: () {
                   showDialog(
                       context: context,
-                      builder: (_) => ImageDialog(image: snapshot.data));
+                      builder: (_) => ImageDialog(image: snapshot.data!));
                 },
                 child: Container(
                   clipBehavior: Clip.hardEdge,
