@@ -3,19 +3,19 @@ import 'package:flutter/material.dart';
 class RoundedButton extends StatelessWidget {
   final String text;
   final Color color;
-  final Function onTap;
+  final Function? onTap;
   final double minWidth;
   final Color textColor;
   final double fontSize;
   final EdgeInsets padding;
   RoundedButton(
-      {this.text,
-      this.color,
+      {this.text = "",
+      this.color = Colors.green,
       this.onTap,
-      this.minWidth,
-      this.textColor,
-      this.fontSize,
-      this.padding});
+      this.minWidth = 20,
+      this.textColor = Colors.black,
+      this.fontSize = 14,
+      this.padding = const EdgeInsets.all(8)});
 
   @override
   Widget build(BuildContext context) {
@@ -25,10 +25,10 @@ class RoundedButton extends StatelessWidget {
       child: ButtonTheme(
         height: 60.0,
         child: ElevatedButton(
-          child: Text(this.text,
-              style: TextStyle(fontSize: fontSize == null ? 21.0 : fontSize)),
+          child: Text(this.text, style: TextStyle(fontSize: fontSize)),
           onPressed: () {
-            this.onTap();
+            if (this.onTap == null) return;
+            this.onTap!();
           },
           style: ElevatedButton.styleFrom(
             minimumSize: Size(100.0, 50.0),
