@@ -1,5 +1,6 @@
 import 'package:zimple/model/event_layout.dart';
 import 'package:flutter/material.dart';
+import 'package:zimple/widgets/person_circle_avatar.dart';
 import '../model/event.dart';
 
 class EventContainer extends StatelessWidget {
@@ -7,9 +8,14 @@ class EventContainer extends StatelessWidget {
   final Event event;
   final EventLayout eventLayout;
   final Function(Event) didTapEvent;
-  EventContainer({this.event, this.eventLayout, this.didTapEvent});
+  EventContainer(
+      {required this.event,
+      required this.eventLayout,
+      required this.didTapEvent});
   @override
   Widget build(BuildContext context) {
+    var isEventLarge = (eventLayout.width > 120 && eventLayout.height > 200);
+    var isEventTextable = (eventLayout.width > 18 && eventLayout.height > 30);
     return Padding(
       padding: EdgeInsets.only(right: 1.0),
       child: Container(
@@ -21,16 +27,31 @@ class EventContainer extends StatelessWidget {
           color: event.color.withAlpha(190),
           borderRadius: BorderRadius.circular(3.0),
         ),
-        child: GestureDetector(
+        child: InkWell(
+          borderRadius: BorderRadius.all(Radius.circular(12.0)),
+          splashColor: Colors.white.withAlpha(190),
           onTap: () {
             didTapEvent(event);
           },
-          onLongPress: () {
-            print("Long pressed event");
-          },
-          child: (eventLayout.width > 18 && eventLayout.height > 30)
-              ? Text(event.title,
-                  style: TextStyle(fontWeight: FontWeight.w300, fontSize: 11.0))
+          child: isEventTextable
+              ? Padding(
+                  padding: isEventLarge ? EdgeInsets.all(6.0) : EdgeInsets.zero,
+                  child: Column(
+                    crossAxisAlignment: CrossAxisAlignment.start,
+                    children: [
+                      Text(event.title,
+                          style: TextStyle(
+                              fontWeight: FontWeight.w300,
+                              fontSize: isEventLarge ? 17 : 11)),
+                      isEventLarge ? SizedBox(height: 8.0) : Container(),
+                      isEventLarge ? Text(event.customer ?? "") : Container(),
+                      isEventLarge ? SizedBox(height: 8.0) : Container(),
+                      isEventLarge
+                          ? ListPersonCircleAvatar(persons: event.persons ?? [])
+                          : Container()
+                    ],
+                  ),
+                )
               : null,
         ),
       ),
