@@ -1,7 +1,11 @@
 import 'package:flutter/material.dart';
 
 class EventLayout {
-  EventLayout({this.height, this.width, this.top, this.left});
+  EventLayout(
+      {required this.height,
+      required this.width,
+      required this.top,
+      required this.left});
   double height;
   double width;
   double top;
