@@ -2,6 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:persistent_bottom_nav_bar/persistent-tab-view.dart';
 import 'package:provider/provider.dart';
 import 'package:zimple/components/timeplan.dart';
+import 'package:zimple/model/event.dart';
 import 'package:zimple/model/user_parameters.dart';
 import 'package:zimple/managers/event_manager.dart';
 import 'package:zimple/utils/constants.dart';
@@ -14,7 +15,9 @@ import 'add_timereport_screen.dart';
 class TimeReportingSelectScreen extends StatefulWidget {
   static const routeName = "time_reporting_select_screen";
   final EventManager eventManager;
-  TimeReportingSelectScreen({this.eventManager});
+  final Function(Event) didSelectEvent;
+  TimeReportingSelectScreen(
+      {required this.eventManager, required this.didSelectEvent});
 
   @override
   _TimeReportingSelectScreenState createState() =>
@@ -22,20 +25,23 @@ class TimeReportingSelectScreen extends StatefulWidget {
 }
 
 class _TimeReportingSelectScreenState extends State<TimeReportingSelectScreen> {
-  EventManager _filteredEventManager(BuildContext context) {
-    UserParameters user =
-        Provider.of<ManagerProvider>(context, listen: false).user;
-    var eventManager = widget.eventManager;
-    eventManager.eventFilter = (events) {
-      return events.where((event) {
-        return event.persons.map((p) => p.id).contains(user.token);
-      }).toList();
-    };
-    return eventManager;
-  }
+  // EventManager _filteredEventManager(BuildContext context) {
+  //   UserParameters user =
+  //       Provider.of<ManagerProvider>(context, listen: false).user;
+  //   var eventManager = widget.eventManager;
+  //   eventManager.eventFilter = (events) {
+  //     return events.where((event) {
+  //       return event.persons.map((p) => p.id).contains(user.token);
+  //     }).toList();
+  //   };
+  //   return eventManager;
+  // }
 
   @override
   Widget build(BuildContext context) {
+    UserParameters user =
+        Provider.of<ManagerProvider>(context, listen: true).user;
+
     return Stack(
       children: [
         Scaffold(
@@ -43,7 +49,7 @@ class _TimeReportingSelectScreenState extends State<TimeReportingSelectScreen> {
             backgroundColor: primaryColor,
             elevation: 0.0,
             toolbarHeight: 75,
-            title: Text("Välj vilket event du vill tidrapportera för",
+            title: Text("Välj vilken arbetsorder du vill tidrapportera för",
                 maxLines: 2, style: TextStyle(color: Colors.white)),
             leading: IconButton(
               icon: Icon(Icons.arrow_back, color: Colors.white),
@@ -53,28 +59,15 @@ class _TimeReportingSelectScreenState extends State<TimeReportingSelectScreen> {
           body: Stack(
             children: [
               Timeplan(
-                eventManager: _filteredEventManager(context),
+                eventManager: widget.eventManager,
                 appBarEnabled: false,
                 shouldShowIsTimereported: true,
+                userIdToOnlyShow: user.token,
                 didTapEvent: (event) {
-                  pushNewScreen(context,
-                      screen: AddTimeReportingScreen(
-                        selectedEvent: event,
-                      ));
+                  widget.didSelectEvent(event);
+                  Navigator.of(context).pop();
                 },
               ),
-              Padding(
-                padding: const EdgeInsets.only(bottom: 25.0),
-                child: Align(
-                  alignment: Alignment.bottomCenter,
-                  child: RectangularButton(
-                      onTap: () {
-                        pushNewScreen(context,
-                            screen: AddTimeReportingScreen());
-                      },
-                      text: "Rapportera ej planerad tid"),
-                ),
-              )
             ],
           ),
         )
