@@ -4,7 +4,7 @@ import 'package:zimple/widgets/provider_widget.dart';
 import '../utils/constants.dart';
 
 class AppBarWidget extends StatelessWidget {
-  final Stream<DateTime> dateStream;
+  final Stream<DateTime>? dateStream;
   final String title;
   AppBarWidget({this.dateStream, this.title = ""});
   @override
@@ -16,9 +16,9 @@ class AppBarWidget extends StatelessWidget {
           return AppBar(
             title: dateStream != null
                 ? Text(
-                    dateStringMonth(snapshot.data) +
+                    dateStringMonth(snapshot.data!) +
                         "  |  V." +
-                        weekNumber(snapshot.data).toString(),
+                        weekNumber(snapshot.data!).toString(),
                     style: TextStyle(
                         color: Colors.white,
                         fontSize: 25.0,
@@ -27,12 +27,12 @@ class AppBarWidget extends StatelessWidget {
             leading: IconButton(
               icon: Icon(Icons.menu),
               onPressed: () {
-                ProviderWidget.of(context).drawerKey.currentState.openDrawer();
+                ProviderWidget.of(context).drawerKey.currentState?.openDrawer();
               },
             ),
             backgroundColor: primaryColor,
             toolbarHeight: 75.0,
-            iconTheme: IconThemeData(color: Colors.white),
+            //iconTheme: IconThemeData(color: Colors.white),
             elevation: 0.0,
           );
         });
