@@ -4,7 +4,10 @@ class ConditionalWidget extends StatelessWidget {
   final Widget childTrue;
   final Widget childFalse;
   final bool condition;
-  ConditionalWidget({this.condition, this.childTrue, this.childFalse});
+  ConditionalWidget(
+      {required this.condition,
+      required this.childTrue,
+      required this.childFalse});
   @override
   Widget build(BuildContext context) {
     return condition ? childTrue : childFalse;
