@@ -3,11 +3,14 @@ import 'package:zimple/model/person.dart';
 
 class PersonCircleAvatar extends StatelessWidget {
   final Person person;
-  PersonCircleAvatar({this.person});
+  final double radius;
+  final double fontSize;
+  PersonCircleAvatar(
+      {required this.person, this.radius = 15, this.fontSize = 11.0});
   @override
   Widget build(BuildContext context) {
     return CircleAvatar(
-      radius: 15,
+      radius: this.radius,
       backgroundColor: Colors.grey.shade400,
       child: Center(
         child: Text(
@@ -17,5 +20,33 @@ class PersonCircleAvatar extends StatelessWidget {
         ),
       ),
     );
+  }
+}
+
+class ListPersonCircleAvatar extends StatelessWidget {
+  final List<Person> persons;
+  final double spacing;
+  final double radius;
+  final double fontSize;
+  const ListPersonCircleAvatar({
+    required this.persons,
+    this.spacing = 5.0,
+    this.radius = 10,
+    this.fontSize = 11.0,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return Row(
+        children: List.generate(persons.length, (index) {
+      var person = persons[index];
+      return Row(
+        children: [
+          PersonCircleAvatar(
+              person: person, radius: this.radius, fontSize: this.fontSize),
+          SizedBox(width: spacing)
+        ],
+      );
+    }));
   }
 }
